requirements.txt
# Lab Intelligence Chatbot - Main Requirements
# Install this for backend development

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# LLM and AI
langchain==0.1.7
langchain-openai==0.0.5
openai==1.12.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging and monitoring
python-json-logger==2.0.7

